update int to float

+++ test/integration/global_api/test_price_and_carbon_intensity.py
@@ -26,7 +26,7 @@ def test_fetch_carbon_intensity_pt():
 
     assert len(result) > 0
     assert isinstance(result[0][0], datetime)
-    assert isinstance(result[0][1], int)
+    assert isinstance(result[0][1], float)
 
 
 @pytest.mark.integration
@@ -39,4 +39,4 @@ def test_fetch_carbon_intensity_dk():
 
     assert len(result) > 0
     assert isinstance(result[0][0], datetime)
-    assert isinstance(result[0][1], int)
+    assert isinstance(result[0][1], float)
